+++ packages/ibm-products/src/components/PageHeader/next/PageHeader.tsx
@@ -771,7 +771,7 @@ const PageHeaderTabBar = React.forwardRef<
   if (!tags) {
     return (
       <div className={classNames} ref={ref} {...other}>
-        <Grid>
+        <Grid condensed>
           <Column lg={16} md={8} sm={4}>
             {children}
             {renderScroller()}
@@ -783,7 +783,7 @@ const PageHeaderTabBar = React.forwardRef<
 
   return (
     <div className={classNames} ref={ref} {...other}>
-      <Grid>
+      <Grid condensed>
         <Column lg={16} md={8} sm={4}>
           <div
             className={classnames(`${blockClass}__tab-bar--tablist`, {
@@ -818,11 +818,12 @@ interface PageHeaderTagOverflowProps {
 const PageHeaderTagOverflow = React.forwardRef<
   HTMLDivElement,
   PageHeaderTagOverflowProps
->(({ renderOverflowTag, renderPopoverContent, children }) => {
+>(({ renderOverflowTag, renderPopoverContent, children }, ref) => {
   const [openPopover, setOpenPopover] = useState(false);
   const [hiddenTags, setHiddenTags] = useState<HTMLElement[]>([]);
 
-  const tagsContainerRef = useRef<HTMLDivElement>(null);
+  const localRef = useRef<HTMLDivElement>(null);
+  const tagsContainerRef = (ref || localRef) as RefObject<HTMLDivElement>;
   // To close popover when window resizes
   useEffect(() => {
     const handleResize = () => {
@@ -848,6 +849,8 @@ const PageHeaderTagOverflow = React.forwardRef<
         setHiddenTags(hidden);
       },
     });
+    // Don't want ref in dependency array
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (